Fix warnings from not having "include" as SYSTEM (cmake)
Summary: No changes to cmake, just some fixes in advance for them.

Test Plan: All tests pass

Reviewers: #poprithms, evan

Reviewed By: #poprithms, evan

Maniphest Tasks: T20625

Differential Revision: https://phabricator.sourcevertex.net/D25805

+++ poprithms/poprithms/include/poprithms/schedule/anneal/graph.hpp
@@ -223,7 +223,7 @@ public:
   AllocWeight getSumLiveness() const;
 
   AllocWeight scheduleToLiveness(ScheduleIndex i) const {
-    return schToLiveness[i];
+    return schToLiveness[static_cast<uint64_t>(i)];
   }
   OpAddress scheduleToOp(ScheduleIndex i) const {
     return schToOp[static_cast<uint64_t>(i)];
@@ -243,7 +243,7 @@ public:
 
   // the allocs required by the op at a schedule index
   const std::vector<AllocAddress> &scheduleToAllocs(ScheduleIndex i) const {
-    return schToAllocs[i];
+    return schToAllocs[static_cast<uint64_t>(i)];
   }
 
   // schedule indices of an ops inputs, sorted
